Split out the actions
